Add tests for the Arden Syntax Version 2.5 specification
This adds a lot of tests, which are derived form the Arden Syntax Version 2.5 specification. Most of the statements made in the specification are backed by a test. This should lead to a very high test coverage.

The test structure and names (packages, classes, methods) are closely related to the chapters of the specification document. This should make it easier to look up the original statement on which the test is based or add more tests.

This test collection is compiler independent. To test a different Arden Syntax compiler, the TestCompiler interface has to be implemented and put as the compiler for the base test class (CompilerTest) which all tests extend. The output has to be filtered to match the requirements of TestCompilerResult.

The ArdenCodeBuilder can be used to easily create code and keep tests short. For custom asserts the base test class (CompilerTest) can be used.