Add event engine
The event engine waits for events and calls and invokes MLMs when they are
scheduled. It manages delayed calls (e.g. in the actions slot) and periodic
or delayed triggers (in the evoke slot). MLMs which are waiting for execution
are called in the order of their priority or their urgency (action slot
calls).

Threads can communicate with its scheduling loop via a message queue.
Messages are EventCalls or MlmCalls. EventCalls are handled
first, then MlmCalls in order of their priority/urgency.