Add bash run script
Ant is not suitable for a run task as it has problems with the arden2bytecode args, so the run task is removed.
Args can be given from the command line as an ant property via -D<property=... or via the <Input/> task.
But quotes (e.g. in arden string constant args) are silently removed by ant.
Escaping can only be done via &quot; in xml-files, but not via the command line.
A bash script also allows piping input/output and doesn't pollute the output with ant messages.