Add ArdenEvent which stores its eventtime
The 'TIME OF an_event' (the primary time) is the clinically relevant time, e.g. the time when a sample was taken.
The 'EVENTTIME' is the time when the relevant data, e.g. the samples measured values, are stored in the database.
This allows using the 'EVENTTIME', to get the matching database entry, while 'TIME OF an_event' can be used for calculation.

Therefore the primary time and eventtime are stored separately.