Add createCodeBuilder() method that initializes an ArdenCodeBuilder with the correct version
The method initializes a code builder with the "arden:" slot set to the highest possible version, restricted by the compiler's target version or by the tests max compatible version.
For versions greater or equal to version 2.9 the resources category is used in the builder's template, as it is no longer optional.
For version 1 the "mlmname:" slot is replace by the "filename:" slot and the "arden:" slot is removed.

This allows tests to influence which template is choosen, via the Compatibility annotation.

createEmptyLogicSlotCodeBuilder() creates a template without the "CONCLUDE TRUE" statement in the logic slot.