Add methods findModules and callEventWithDelay to execution contexts
The method findModules(ArdenEvent) allows finding MLMs that should normally evoked for an event.
This will be used for event calls in the logic slot, which returns the return values of all MLMs that are evoked by the event.

The method callEventWithDelay(ArdenEvent, ArdenValue) will be used for action slot event calls.