Split INTERFACE and MLM CALL logic
MLM variables can be used in the INCLUDE statement, whereas INTERFACE variables can not be used.
ExecutionContext#findModule() must therefore return an MLM instead of an ArdenRunnable, to differentiate these two kinds of variables.

It is currently possible to use INTERFACES in action Slot CALL statements. The specification only mentions that MLMs and EVENTs can be called.
This split allows INTERFACES to be handled differently, should this actually be the case.

An MLM can still be implemented as pure Java Code, by subclassing MedicalLogicModule instead of ArdenRunnable.