Change CyclicTrigger to always return the next cycles runtime
All cycles of the trigger are returned, even if they are in the past.
It is up to the context/engine to decide, if past triggers should run.
This may make it easier implement a way to persist these triggers, between restarts and e.g. call them via cron jobs.
It also fixes some nasty of by one millisecond errors :)